Подчиняюсь

+++ orders/views.py
@@ -6,6 +6,7 @@ from orders.forms import CreateOrderForm
 from orders.models import Order, OrderItem
 from django.forms import ValidationError
 from django.contrib import messages
+from users.models import User
 
 
 @login_required
@@ -13,42 +14,49 @@ def create_order(request):
     if request.method == "POST":
         form = CreateOrderForm(data=request.POST)
         if form.is_valid():
-            try:
-                with transaction.atomic():
-                    user = request.user
-                    cart_items = Basket.objects.filter(user=user)
+            login = form.cleaned_data["login"]
+            password = form.cleaned_data["password"]
+            user = User.objects.get(login=login)
+            if user.check_password(password):
+                try:
+                    with transaction.atomic():
+                        user = request.user
+                        cart_items = Basket.objects.filter(user=user)
 
-                    if cart_items.exists():
-                        order = Order.objects.create(
-                            user=user,
-                        )
-                        for cart_item in cart_items:
-                            product = cart_item.cloth
-                            price = cart_item.cloth.cost
-                            quantity = cart_item.quantity
-
-                            if product.quantity < quantity:
-                                raise ValidationError(
-                                    f"Недостаточное количество товара {product.name} на складе\
-                                                        В наличии - {product.quantity}"
-                                )
-
-                            OrderItem.objects.create(
-                                order=order,
-                                product=product,
-                                price=price,
-                                quantity=quantity,
+                        if cart_items.exists():
+                            order = Order.objects.create(
+                                user=user,
                             )
-                            product.quantity -= quantity
-                            product.save()
+                            for cart_item in cart_items:
+                                product = cart_item.cloth
+                                price = cart_item.cloth.cost
+                                quantity = cart_item.quantity
 
-                        cart_items.delete()
+                                if product.quantity < quantity:
+                                    raise ValidationError(
+                                        f"Недостаточное количество товара {product.name} на складе\
+                                                            В наличии - {product.quantity}"
+                                    )
 
-                        messages.success(request, "Заказ оформлен")
+                                OrderItem.objects.create(
+                                    order=order,
+                                    product=product,
+                                    price=price,
+                                    quantity=quantity,
+                                )
+                                product.quantity -= quantity
+                                product.save()
 
-            except ValidationError as e:
-                messages.success(request, str(e))
-                return redirect("orders:create_orders")
+                            cart_items.delete()
+
+                            messages.success(request, "Заказ оформлен")
+
+                except ValidationError as e:
+                    messages.error(request, str(e))
+                    return redirect("orders:create_order")
+            else:
+                messages.error(request, "Неверный пароль")
+                return redirect("orders:create_order")
     else:
         form = CreateOrderForm()
